# shared/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AERO API",
        "description": "API do ewidencji operacji lotniczych i zlecen na lot helikopterem",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "Bearer": []
        }
    ],
    "paths": {
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/landing-sites/{id}": {
            "get": {
                "tags": [
                    "landing-site-controller"
                ],
                "operationId": "findById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LandingSiteResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "landing-site-controller"
                ],
                "operationId": "update_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLandingSiteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LandingSiteResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/helicopters/{id}": {
            "get": {
                "tags": [
                    "helicopter-controller"
                ],
                "operationId": "findById_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HelicopterResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "helicopter-controller"
                ],
                "operationId": "update_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateHelicopterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/HelicopterResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/flight-orders/{id}": {
            "get": {
                "tags": [
                    "flight-order-controller"
                ],
                "operationId": "findById_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlightOrderResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "flight-order-controller"
                ],
                "operationId": "update_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFlightOrderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlightOrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/flight-operations/{id}": {
            "get": {
                "tags": [
                    "flight-operation-controller"
                ],
                "operationId": "findById_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlightOperationResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "flight-operation-controller"
                ],
                "operationId": "update_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFlightOperationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FlightOperationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/crew-members/{id}": {
            "get": {
                "tags": [
                    "crew-member-controller"
                ],
                "operationId": "findById_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CrewMemberResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "crew-member-controller"
                ],
                "operationId": "update_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCrewMemberRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CrewMemberResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/landing-sites": {
            "get": {
                "tags": [
                    "landing-site-controller"
                ],
                "operationId": "findAll_1",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LandingSiteResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "landing-site-controller"
                ],
                "operationId": "create_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLandingSiteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LandingSiteResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/crew-members": {
            "get": {
                "tags": [
                    "crew-member-controller"
                ],
                "operationId": "findAll_5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CrewMemberResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "crew-member-controller"
                ],
                "operationId": "create_5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCrewMemberRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CrewMemberResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UpdateUserRequest": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "lastName": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 8
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMINISTRATOR",
                            "PLANNER",
                            "SUPERVISOR",
                            "PILOT"
                        ]
                    }
                },
                "required": [
                    "email",
                    "firstName",
                    "lastName",
                    "role"
                ]
            },
            "UserResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "ADMINISTRATOR",
                            "PLANNER",
                            "SUPERVISOR",
                            "PILOT"
                        ]
                    }
                },
                "required": [
                    "email",
                    "firstName",
                    "id",
                    "lastName",
                    "role"
                ]
            },
            "UpdateLandingSiteRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "latitude",
                    "longitude",
                    "name"
                ]
            },
            "LandingSiteResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "id",
                    "latitude",
                    "longitude",
                    "name"
                ]
            },
            "UpdateHelicopterRequest": {
                "type": "object",
                "properties": {
                    "registrationNumber": {
                        "type": "string",
                        "maxLength": 30,
                        "minLength": 0
                    },
                    "helicopterType": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "maxCrewCount": {
                        "type": "integer",
                        "format": "int32",
                        "maximum": 10,
                        "minimum": 1
                    },
                    "maxCrewWeight": {
                        "type": "integer",
                        "format": "int32",
                        "maximum": 1000,
                        "minimum": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "INACTIVE"
                        ]
                    },
                    "inspectionExpiryDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "rangeKm": {
                        "type": "integer",
                        "format": "int32",
                        "maximum": 1000,
                        "minimum": 1
                    }
                },
                "required": [
                    "helicopterType",
                    "maxCrewCount",
                    "maxCrewWeight",
                    "rangeKm",
                    "registrationNumber",
                    "status"
                ]
            },
            "HelicopterResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "registrationNumber": {
                        "type": "string"
                    },
                    "helicopterType": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "maxCrewCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "maxCrewWeight": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "INACTIVE"
                        ]
                    },
                    "inspectionExpiryDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "rangeKm": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "required": [
                    "helicopterType",
                    "id",
                    "maxCrewCount",
                    "maxCrewWeight",
                    "rangeKm",
                    "registrationNumber",
                    "status"
                ]
            },
            "UpdateFlightOrderRequest": {
                "type": "object",
                "properties": {
                    "plannedStartTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "plannedEndTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "helicopterId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "crewMemberIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    },
                    "departureSiteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "arrivalSiteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "operationIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    },
                    "estimatedRouteLengthKm": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "actualStartTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "actualEndTime": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "arrivalSiteId",
                    "crewMemberIds",
                    "departureSiteId",
                    "estimatedRouteLengthKm",
                    "helicopterId",
                    "operationIds",
                    "plannedEndTime",
                    "plannedStartTime"
                ]
            },
            "FlightOrderResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "plannedStartTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "plannedEndTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "pilotId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "INTRODUCED",
                            "SUBMITTED",
                            "REJECTED",
                            "ACCEPTED",
                            "PARTIALLY_COMPLETED",
                            "COMPLETED",
                            "NOT_COMPLETED"
                        ]
                    },
                    "helicopterId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "crewMemberIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    },
                    "crewWeight": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "departureSiteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "arrivalSiteId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "operationIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    },
                    "estimatedRouteLengthKm": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "actualStartTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "actualEndTime": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "arrivalSiteId",
                    "createdAt",
                    "crewMemberIds",
                    "crewWeight",
                    "departureSiteId",
                    "estimatedRouteLengthKm",
                    "helicopterId",
                    "id",
                    "operationIds",
                    "pilotId",
                    "plannedEndTime",
                    "plannedStartTime",
                    "status",
                    "updatedAt"
                ]
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}